[d3-chart] Rename Chart to Plot
