+++ semcore/d3-chart/src/style/chart.shadow.css
@@ -1,4 +1,4 @@
 @import '@semcore/utils/style/var.css';
 
-SChart {
+SPlot {
 }
