+++ semcore/d3-chart/src/types/Chart.ts
@@ -2,7 +2,7 @@ import { IBoxProps } from '@semcore/flex-box';
 import { CProps, ReturnEl } from '@semcore/core';
 import IContext from './context';
 
-export interface IChartProps extends IContext, IBoxProps {
+export interface IPlotProps extends IContext, IBoxProps {
   /** Width svg element
    * @default 0*/
   width?: number;
@@ -11,6 +11,6 @@ export interface IChartProps extends IContext, IBoxProps {
   height?: number;
 }
 
-declare const Chart: <T>(props: CProps<IChartProps & T>) => ReturnEl;
+declare const Plot: <T>(props: CProps<IPlotProps & T>) => ReturnEl;
 
-export default Chart;
+export default Plot;
