+++ website/docs/data-display/d3-chart/d3-chart-api.md
@@ -2,7 +2,7 @@
 title: API
 ---
 
-@## Chart
+@## Plot
 
 Root element for all charts.
