+++ website/docs/data-display/d3-chart/examples/bar-horizontal.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Chart, HorizontalBar, YAxis, XAxis, HoverRect, Tooltip } from '@semcore/d3-chart';
+import { Plot, HorizontalBar, YAxis, XAxis, HoverRect, Tooltip } from '@semcore/d3-chart';
 import { scaleLinear, scaleBand } from 'd3-scale';
 import { Flex } from '@semcore/flex-box';
 import { Text } from '@semcore/typography';
@@ -20,7 +20,7 @@ export default () => {
     .paddingOuter(0.2);
 
   return (
-    <Chart data={data} scale={[xScale, yScale]} width={width} height={height}>
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
       <YAxis hide={false} ticks={yScale.domain()}>
         <YAxis.Ticks />
       </YAxis>
@@ -44,7 +44,7 @@ export default () => {
         }}
       </Tooltip>
       <HorizontalBar x="bar" y="category" />
-    </Chart>
+    </Plot>
   );
 };
