+++ website/docs/data-display/d3-chart/examples/donut-without-data.js
@@ -1,16 +1,16 @@
 import React from 'react';
-import { Chart, Donut } from '@semcore/d3-chart';
+import { Plot, Donut } from '@semcore/d3-chart';
 
 export default () => {
   return (
-    <Chart width={300} height={300} data={data}>
+    <Plot width={300} height={300} data={data}>
       <Donut innerRadius={100}>
         <Donut.NotData />
         <Donut.Pie dataKey="a" />
         <Donut.Pie dataKey="b" color="#3AB011" />
         <Donut.Pie dataKey="c" color="#890C85" />
       </Donut>
-    </Chart>
+    </Plot>
   );
 };
