+++ website/docs/data-display/d3-chart/examples/margin.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Chart, Line, minMax } from '@semcore/d3-chart';
+import { Plot, Line, minMax } from '@semcore/d3-chart';
 import { scaleLinear } from 'd3-scale';
 
 export default () => {
@@ -16,7 +16,7 @@ export default () => {
     .domain(minMax(data, 'y'));
 
   return (
-    <Chart
+    <Plot
       data={data}
       scale={[xScale, yScale]}
       width={width}
@@ -24,7 +24,7 @@ export default () => {
       style={{ border: '1px solid' }}
     >
       <Line x="x" y="y" />
-    </Chart>
+    </Plot>
   );
 };
