+++ website/docs/data-display/d3-chart/examples/semi-donut.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import { Chart, Donut, Tooltip } from '@semcore/d3-chart';
+import { Plot, Donut, Tooltip } from '@semcore/d3-chart';
 import { Text } from '@semcore/typography';
 import { Flex } from '@semcore/flex-box';
 
 export default () => {
   return (
-    <Chart width={300} height={150} data={data}>
+    <Plot width={300} height={150} data={data}>
       <Donut half innerRadius={100}>
         <Donut.Pie dataKey="a" name="Pie 1" />
         <Donut.Pie dataKey="b" color="#890C85" name="Pie 2" />
@@ -33,7 +33,7 @@ export default () => {
           }}
         </Tooltip>
       </Donut>
-    </Chart>
+    </Plot>
   );
 };
