+++ website/docs/data-display/d3-chart/examples/stacked-area.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-  Chart,
+  Plot,
   XAxis,
   YAxis,
   minMax,
@@ -32,7 +32,7 @@ export default () => {
     .domain([0, 15]);
 
   return (
-    <Chart data={data} scale={[xScale, yScale]} width={width} height={height}>
+    <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>
       <YAxis ticks={yScale.ticks()}>
         <YAxis.Ticks />
         <YAxis.Grid />
@@ -97,7 +97,7 @@ export default () => {
           <StackedArea.Area.Dots />
         </StackedArea.Area>
       </StackedArea>
-    </Chart>
+    </Plot>
   );
 };
